switchover for citizen and delete unused frontend code
