--- frontend/src/citizen-frontend/children/sections/vasu-and-leops/vasu/vasu-events.ts
@@ -1,19 +1,0 @@
-// SPDX-FileCopyrightText: 2017-2022 City of Espoo
-//
-// SPDX-License-Identifier: LGPL-2.1-or-later
-
-import { VasuDocumentEvent } from 'lib-common/generated/api-types/vasu'
-import HelsinkiDateTime from 'lib-common/helsinki-date-time'
-
-export const getLastPublished = (
-  events: VasuDocumentEvent[]
-): HelsinkiDateTime | undefined =>
-  events.reduce<HelsinkiDateTime | undefined>((acc, event) => {
-    if (event.eventType !== 'PUBLISHED') {
-      return acc
-    }
-    if (!acc) {
-      return event.created
-    }
-    return event.created > acc ? event.created : acc
-  }, undefined)
